Sistemato l'errore sulla generazione dei pianeti Modificati gli eventi riguardanti i pulsanti premuti, adesso più movimenti possono occorrere contemporaneamente

+++ Project/Project/pianeta.cpp
@@ -11,7 +11,7 @@ Pianeta::Pianeta(int id, float cord_x, float cord_y, unsigned int width, unsigne
 }
 
 int Pianeta::getRaggio() {
-	return shape_.getRadius();
+	return (int) shape_.getRadius();
 }
 
 sf::Vector2f Pianeta::getPosizione()
